Update Holacracy-Constitution.md
It seems to me confusing not to refer to the work of the Circle that includes the Role. Role is not free electrons and this would harm the holarchy. This question surface quite often and this is not so obvious at least at the earlier stages that a Role is doing part of the Circle containing it's work. I don't really know how to put it in correct English, I give it a try.