Circle Supporters
The issue is that the current mental model re role/circle is completely changed from the current 4.1 we are used to. For instance, from a position of a Role inside the Circle, take the Role Supporter ("on the Circle") concept defined at section 1.2.3, it is doing work within the Circle like a peer, then how could you reference it from a position of a Role inside the Circle? There is a need to define the Circle Supporters piece which is missing.